Fix not converging to good return
"It's times like these you learn to love *static typing* again".
Went crazy for  a week over a subtle bug caused by torch.Tensor(x)  != torch.tensor(x) and fixed it. This is important to fix as we normalize image based on checking if dtype of observation is uint8.
This single handedly fixes the performance bug as we are normalizing all images (during stepping through the env with our agent policy).

Added proper truncated normal distribution instead of the hacky one. This fixes entropy mismatch with tensorflow.